Primer commit del nuevo proyecto

+++ app/models/liquidacion.py
@@ -9,6 +9,8 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
 from app.models.empleado import Empleado
+from decimal import Decimal
+
 
 
 class Liquidacion(Base):
@@ -46,18 +48,18 @@ class Liquidacion(Base):
 
     # Valor base de la hora utilizado para calcular la liquidacion. Se guarda
     # como importe monetario usando precision decimal.
-    valor_hora_base: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    valor_hora_base: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
 
     # Suma parcial correspondiente al calculo de horas antes de adicionales y
     # descuentos.
     subtotal_horas: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
 
     # Importe adicional generado por las horas trabajadas en feriado.
-    monto_feriado: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    monto_feriado: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
 
     # Monto adicional por asistencia perfecta. Por defecto comienza en cero si
     # no corresponde aplicar este concepto.
-    asistencia_perfecta: Mapped[float] = mapped_column(
+    asistencia_perfecta: Mapped[Decimal] = mapped_column(
         Numeric(12, 2),
         nullable=False,
         default=0,
@@ -65,7 +67,7 @@ class Liquidacion(Base):
 
     # Total de descuentos aplicados a la liquidacion. Por defecto se inicializa
     # en cero hasta que existan descuentos a computar.
-    total_descuentos: Mapped[float] = mapped_column(
+    total_descuentos: Mapped[Decimal] = mapped_column(
         Numeric(12, 2),
         nullable=False,
         default=0,
@@ -73,7 +75,7 @@ class Liquidacion(Base):
 
     # Resultado final neto que percibe el empleado luego de sumar conceptos y
     # restar descuentos.
-    total_neto: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    total_neto: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
 
     # Representacion textual del total neto, util para recibos o reportes.
     total_en_letras: Mapped[str] = mapped_column(String(255), nullable=False)
